feat: added multi-account support for classroom

+++ src/main/kotlin/com/epcc/politech_manager/classroom/ClassroomController.kt
@@ -1,29 +1,77 @@
 package com.epcc.politech_manager.classroom
 
+import com.epcc.politech_manager.error.ExceptionUserModel
+import com.epcc.politech_manager.error.UserException
+import com.epcc.politech_manager.user.UserEntityDAO
+import com.epcc.politech_manager.user.UserService
+import com.epcc.politech_manager.utils.ResponseOk
 import org.springframework.web.bind.annotation.*
 
 @RestController
-class ClassroomController(val service: ClassroomService) {
+class ClassroomController(val service: ClassroomService, val userService: UserService) {
     @GetMapping("/classroom")
-    fun index(): List<ClassroomEntity> = service.getAllClassrooms()
+    fun index(@RequestHeader("Authorization") auth: String): List<ClassroomEntityDTO> {
+        val user: UserEntityDAO? = userService.getUserWithToken(auth)
+        if (user != null) {
+            return service.getAllClassrooms().filter { it.user == user }.map { it.toDTO() }
+        } else {
+            throw UserException(ExceptionUserModel.WRONG_USER)
+        }
+    }
 
     @PostMapping("/classroom")
-    fun post(@RequestBody classroom: ClassroomEntity) {
-        service.post(classroom)
+    fun post(@RequestHeader("Authorization") auth: String,
+             @RequestBody classroom: ClassroomEntityDTO)
+    : ResponseOk {
+        val user: UserEntityDAO? = userService.getUserWithToken(auth)
+        if (user != null) {
+            service.post(classroom.toDAO(user))
+            return ResponseOk(200, "Classroom successfully created")
+        } else {
+            throw UserException(ExceptionUserModel.WRONG_USER)
+        }
     }
 
     @GetMapping("/classroom/{id}")
-    fun getClassroom(@PathVariable id: Long): ClassroomEntity? {
-        return service.getClassroom(id)
+    fun getClassroom(@RequestHeader("Authorization") auth: String,
+                     @PathVariable id: Long)
+    : ClassroomEntityDTO? {
+        val user: UserEntityDAO? = userService.getUserWithToken(auth)
+        if (user != null) {
+            val department = service.getClassroom(id)
+            if (department?.user == user) {
+                return department.toDTO()
+            } else {
+                throw UserException(ExceptionUserModel.WRONG_USER)
+            }
+        } else {
+            throw UserException(ExceptionUserModel.WRONG_USER)
+        }
     }
 
     @PostMapping("/classroom/delete/{id}")
-    fun deleteClassroom(@PathVariable id: Long) {
-        service.deleteClassroom(id)
+    fun deleteClassroom(@RequestHeader("Authorization") auth: String,
+                        @PathVariable id: Long)
+    : ResponseOk {
+        val user: UserEntityDAO? = userService.getUserWithToken(auth)
+        if (user != null) {
+            service.deleteClassroom(id)
+            return ResponseOk(200,"Classroom successfully deleted")
+        } else {
+            throw UserException(ExceptionUserModel.WRONG_USER)
+        }
     }
 
     @PostMapping("/classroom/update")
-    fun updateClassroom(@RequestBody classroom: ClassroomEntity) {
-        service.updateClassroom(classroom)
+    fun updateClassroom(@RequestHeader("Authorization") auth: String,
+                        @RequestBody classroom: ClassroomEntityDTO)
+    : ResponseOk {
+        val user: UserEntityDAO? = userService.getUserWithToken(auth)
+        if (user != null) {
+            service.updateClassroom(classroom.toDAO(user))
+            return ResponseOk(200,"Degree successfully updated")
+        } else {
+            throw UserException(ExceptionUserModel.WRONG_USER)
+        }
     }
 }
